+++ app/domains/fluggastrechte/formular/prozessfuehrung/doneFunctions.ts
@@ -1,6 +1,6 @@
 import type { GenericGuard } from "~/domains/guards.server";
 import { objectKeysNonEmpty } from "~/util/objectKeysNonEmpty";
-import type { FluggastrechteProzessfuehrungContext } from "./context";
+import type { FluggastrechteProzessfuehrungContext } from "./userData";
 
 export const prozessfuehrungDone: GenericGuard<
   FluggastrechteProzessfuehrungContext
