--- app/domains/fluggastrechte/formular/streitwertKosten/context.ts
@@ -1,8 +1,0 @@
-import { z } from "zod";
-import { YesNoAnswer } from "~/services/validation/YesNoAnswer";
-
-export const fluggastrechteStreitKostenDaten = {
-  prozesszinsen: YesNoAnswer,
-};
-
-const _contextObject = z.object(fluggastrechteStreitKostenDaten).partial();
