+++ app/domains/fluggastrechte/formular/streitwertKosten/userData.ts
@@ -1,0 +1,5 @@
+import { YesNoAnswer } from "~/services/validation/YesNoAnswer";
+
+export const fluggastrechteStreitKostenInputSchema = {
+  prozesszinsen: YesNoAnswer,
+};
